refactor: centralized cron job function and load in loader function
